requirements_radu.txt
amqp==2.3.2
attrs==19.1.0
backports-abc==0.5
backports.csv==1.0.7
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.shutil-which==3.5.2
billiard==3.5.0.4
bleach==3.1.0
boto==2.49.0
boto3==1.9.182
botocore==1.12.182
bz2file==0.98
cachetools==3.1.1
celery==4.2.1
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
configparser==3.7.4
cycler==0.10.0
decorator==4.3.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11.16
django-extensions==2.1.3
django-grappelli==2.12.1
django-import-export==1.2.0
django-ipware==2.1.0
django-jfu==2.0.9
django-markdown2==0.3.1
django-model-utils==3.1.2
docutils==0.14
entrypoints==0.3
enum34==1.1.6
et-xmlfile==1.0.1
functools32==3.2.3.post2
futures==3.2.0
gensim==3.7.3
idna==2.8
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
jdcal==1.4.1
Jinja2==2.10.1
jmespath==0.9.4
jsonpickle==1.0
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.5.0
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kombu==4.2.1
markdown2==2.3.6
MarkupSafe==1.1.1
matplotlib==2.2.4
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.8
numpy==1.15.4
odfpy==1.4.0
openpyxl==2.6.2
pandas==0.24.2
pandocfilters==1.4.2
pathlib2==2.3.4
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==1.0.16
psycopg2-binary==2.8.3
Pygments==2.4.2
pymzml==0.7.7
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2018.7
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.2
redis==2.10.6
requests==2.22.0
s3transfer==0.2.1
scandir==1.10.0
scikit-learn==0.20.3
scipy==1.2.2
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
smart-open==1.8.4
tablib==0.13.0
terminado==0.8.2
testpath==0.4.2
tornado==5.1.1
tqdm==4.32.2
traitlets==4.3.2
typing==3.6.6
urllib3==1.25.3
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
win-unicode-console==0.5
xlrd==1.2.0
xlwt==1.3.0
